pyproject.toml
[project]
name = "ai-driven-semantic-log-anomaly-detection"
version = "0.1.0"
description = "AI-driven semantic log anomaly detection with OpenAI embeddings, Qdrant, and HDBSCAN clustering"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22.0",
    "langfuse>=2.0.0",
    "doppler-sdk>=0.1.0",
    "kafka-python>=2.0.2",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "prometheus-client>=0.19.0",
    "ruff>=0.14.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.15",
]

[tool.uv.workspace]
members = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**" = ["ARG"]  # Allow unused arguments in tests

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
